extract MobileSidePanel
