+++ components/Modal.tsx
@@ -1,10 +1,11 @@
 import { Global } from '@emotion/core'
 import { Icon } from '@makerdao/dai-ui-icons'
+import { useSharedUI } from 'components/SharedUIProvider'
 import { ModalProps } from 'helpers/modalHook'
 import { useObservable } from 'helpers/observableHook'
 import { WithChildren } from 'helpers/types'
 import { Trans, useTranslation } from 'next-i18next'
-import React, { useCallback, useEffect, useState } from 'react'
+import React, { ReactNode, useCallback, useEffect, useState } from 'react'
 import ReactDOM from 'react-dom'
 import { TRANSITIONS } from 'theme'
 import {
@@ -235,6 +236,49 @@ export function MobileSidePanelPortal({ children }: WithChildren) {
   const onMobile = useOnMobile()
 
   return onMobile ? ReactDOM.createPortal(children, document.body) : children
+}
+
+export function MobileSidePanel({
+  toggleTitle,
+  children,
+}: WithChildren & { toggleTitle: ReactNode }) {
+  const { vaultFormOpened, setVaultFormOpened, setVaultFormToggleTitle } = useSharedUI()
+
+  useEffect(() => {
+    setVaultFormToggleTitle(toggleTitle)
+
+    return () => {
+      setVaultFormToggleTitle(undefined)
+      setVaultFormOpened(false)
+    }
+  }, [])
+
+  const onClose = () => setVaultFormOpened(false)
+
+  return (
+    <MobileSidePanelPortal>
+      <Box
+        sx={{
+          display: 'block',
+          position: ['fixed', 'relative'],
+          top: 0,
+          left: 0,
+          right: 0,
+          bottom: 0,
+          transition: ['0.3s transform ease-in-out', '0s'],
+          transform: [`translateX(${vaultFormOpened ? '0' : '100'}%)`, 'translateX(0)'],
+          bg: ['background', 'transparent'],
+          p: [3, 0],
+          pt: [0, 0],
+          overflowX: ['hidden', 'visible'],
+          zIndex: ['modal', 0],
+        }}
+      >
+        <MobileSidePanelClose opened={vaultFormOpened} onClose={onClose} />
+        {children}
+      </Box>
+    </MobileSidePanelPortal>
+  )
 }
 
 export function MobileSidePanelClose({
